# Performance improvement in Spark UDFs that return an array